requirements.txt
Flask==0.10.1
Flask-Cache==0.12
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-Migrate==0.1.4
Flask-SQLAlchemy==2.0
Flask-Script==0.6.3
Flask-Security==1.6.9
Flask-UUID==0.2
Flask-WTF==0.13.1
Flask-thumbnails==0.2
flask-paginate==0.2.1
Flask-Principal
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
SQLAlchemy==1.0.6
SQLAlchemy-Utils==0.30.1
Unipath==1.0
WTForms==1.0.5
Werkzeug==0.9.6
alembic==0.6.1
httplib2==0.9
itsdangerous==0.23
nose==1.3.0
pyparsing==2.0.1
python-dateutil==2.2
python-magic==0.4.6
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
Pillow
psycopg2
Flask-Babel
Flask-Admin
sqlalchemy-i18n
markdown
raven
blinker
flask-apidoc
Cerberus<=1.0.1
requests==2.11.1
Celery==4.0.0rc6
kombu==4.0.0
vine==1.1.3
redis==2.10.5
pypyodbc
pycrypto==2.6.1
param==1.5.1
ipython==5.4.1
jupyter==1.0.0
